New Sound cores and fixes
* Fixed flag bug in Z80 core.
* Added new YM-2149 core.
* Added new SN-76489 core.
* Fixed audio mixing to correct signed 16-bit PCM.
* Fixed info-frame data in TMDS core.
* Added Phoenix build system for merging bit-streams with the CV menu code.
* Added missing bin2phx utility and created makefile for it.
* Added missing bit2bin utility and created makefile for it.
* Added missing makemem utility and created makefile for it.
* Included data2mem.exe as part of the included utilities.
* Included the needed SDCC compiler as part of the repo.
* Fixed the clock generation, fixing all failed timing constraints.
* Changed system clock to 25MHz.
* Corrected the 3.579MHz clock-enable generation.
* Fixed wait-state circuit.
* Updated UCF file to support the new LVDS level-shifter.
* Cleaned up most code files for consistency and removed unused signals.
* Updated all comments to ensure accuracy with existing code.
* Tursi updated the CV menu to compile with SDCC 4.0.

Games known to have been fixed by these changes:

  Sound and speech problems fixed:
  * QBIQS
  * Squish'Em Sam
  * Side Trak
  * Children of the Night
  * Wizard of Wor

  Games not previously working, that now work:
  * Sky Jaguar
  * Magical Tree
  * Road Fighter

+++ gameMenus/coleco/src/fatFs/diskio.h
@@ -102,27 +102,27 @@ extern BYTE CardType;			    	/* b0:MMC, b1:SDv1, b2:SDv2, b3:Block addressing */
 ---------------------------------------------------------------------------*/
 
 /* MMC/SD command (SPI mode) */
-#define CMD0	(0)			/* GO_IDLE_STATE */
-#define CMD1	(1)			/* SEND_OP_COND */
-#define	ACMD41	(0x80+41)	/* SEND_OP_COND (SDC) */
-#define CMD8	(8)			/* SEND_IF_COND */
-#define CMD9	(9)			/* SEND_CSD */
-#define CMD10	(10)		/* SEND_CID */
-#define CMD12	(12)		/* STOP_TRANSMISSION */
-#define CMD13	(13)		/* SEND_STATUS */
-#define ACMD13	(0x80+13)	/* SD_STATUS (SDC) */
-#define CMD16	(16)		/* SET_BLOCKLEN */
-#define CMD17	(17)		/* READ_SINGLE_BLOCK */
-#define CMD18	(18)		/* READ_MULTIPLE_BLOCK */
-#define CMD23	(23)		/* SET_BLOCK_COUNT */
-#define	ACMD23	(0x80+23)	/* SET_WR_BLK_ERASE_COUNT (SDC) */
-#define CMD24	(24)		/* WRITE_BLOCK */
-#define CMD25	(25)		/* WRITE_MULTIPLE_BLOCK */
-#define CMD32	(32)		/* ERASE_ER_BLK_START */
-#define CMD33	(33)		/* ERASE_ER_BLK_END */
-#define CMD38	(38)		/* ERASE */
-#define CMD55	(55)		/* APP_CMD */
-#define CMD58	(58)		/* READ_OCR */
+#define CMD0	((BYTE)0)			/* GO_IDLE_STATE */
+#define CMD1	((BYTE)1)			/* SEND_OP_COND */
+#define	ACMD41	((BYTE)(0x80+41))	/* SEND_OP_COND (SDC) */
+#define CMD8	((BYTE)8)			/* SEND_IF_COND */
+#define CMD9	((BYTE)9)			/* SEND_CSD */
+#define CMD10	((BYTE)10)		/* SEND_CID */
+#define CMD12	((BYTE)12)		/* STOP_TRANSMISSION */
+#define CMD13	((BYTE)13)		/* SEND_STATUS */
+#define ACMD13	((BYTE)(0x80+13))	/* SD_STATUS (SDC) */
+#define CMD16	((BYTE)16)		/* SET_BLOCKLEN */
+#define CMD17	((BYTE)17)		/* READ_SINGLE_BLOCK */
+#define CMD18	((BYTE)18)		/* READ_MULTIPLE_BLOCK */
+#define CMD23	((BYTE)23)		/* SET_BLOCK_COUNT */
+#define	ACMD23	((BYTE)(0x80+23))	/* SET_WR_BLK_ERASE_COUNT (SDC) */
+#define CMD24	((BYTE)24)		/* WRITE_BLOCK */
+#define CMD25	((BYTE)25)		/* WRITE_MULTIPLE_BLOCK */
+#define CMD32	((BYTE)32)		/* ERASE_ER_BLK_START */
+#define CMD33	((BYTE)33)		/* ERASE_ER_BLK_END */
+#define CMD38	((BYTE)38)		/* ERASE */
+#define CMD55	((BYTE)55)		/* APP_CMD */
+#define CMD58	((BYTE)58)		/* READ_OCR */
 
 /* MMC card type flags (MMC_GET_TYPE) */
 #define CT_MMC		0x01		/* MMC ver 3 */
